+++ tests/acceptance/graphql/hello.spec.ts
