+++ tests/src/schema/hello-schema.spec.ts
@@ -1,6 +1,6 @@
 import { gql } from 'apollo-server';
 
-import HelloSchema from '../../src/schema/hello-schema';
+import HelloSchema from '../../../src/schema/hello-schema';
 
 describe('Given the HelloSchema', () => {
   test('Then should be expected type', () => {
